Auto-refresh of clusters on team page
* refactoring a cluster list item into it's own component
* this allows it to be controlled individually
* refresh the cluster every 10 seconds until it's status is `Success`
